requirements.txt
langchain==1.1.3
langchain-openai==1.1.3
langchain-community==0.4.1
langchian-text-splitters==1.1.0
langchain-huggingface==1.2.0
faiss-cpu==1.13.1
pandas==2.3.3
pypdf==6.4.2
python-dotenv==1.2.1
ragas==0.4.1
datasets==4.4.1
sentence-transformers==5.2.0
fastapi[all]